backend/requirements.txt
# Backend Dependencies
setuptools>=65.5.0
wheel>=0.37.1
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
pydantic==2.4.2
pydantic-settings==2.0.3
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.0
supabase==2.0.1
bcrypt==4.0.1
email-validator>=2.0.0
# Document processing
pypdf==3.17.0
docx2txt==0.8
pytesseract==0.3.10
# Data processing
numpy>=1.25.0
pandas>=2.0.3
scikit-learn==1.3.0
# Optional ML dependencies that will be used if available
pytorch-tabnet==4.0.0
torch==2.0.1
# Utilities
tqdm==4.66.1
XlsxWriter==3.1.8